fix: move to folder assets and change google logo
