+++ mentesa/src/components/ButtonEnter/index.jsx
@@ -5,7 +5,7 @@ import { useAuth } from '../../hooks/useAuth'
 import { ThemeProvider } from '@mui/material/styles'
 import { themeButtonGoogle } from '../../styles/Theme'
 import Button from '@mui/material/Button'
-import Google from '../../pages/SignIn/images/google.svg'
+import Google from '../../assets/google.png'
 
 export default function GoToGoogle() {
   const { user, signInWithGoogle } = useAuth()
@@ -26,7 +26,7 @@ export default function GoToGoogle() {
   return (
     <ThemeProvider theme={themeButtonGoogle}>
       <Button variant='contained' onClick={handleSignIn}>
-        <img src={Google} alt='logo Google' style={{ marginRight: '10px' }} />
+        <img src={Google} alt='logo Google' style={{ marginRight: '10px', height:'35px', width:'35px' }} />
         Continue com o Google
       </Button>
     </ThemeProvider>
